[part4] finishes 4.15-4.22 - 3.3h

+++ part4/bloglist/controllers/blogs.js
@@ -1,28 +1,46 @@
 const blogsRouter = require("express").Router();
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const userExtractor = require("../utils/middleware").userExtractor;
 
 blogsRouter.get("/", async (_, response) => {
   const blogs = await Blog.find({});
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
-
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id)
+    return response.status(401).json({ error: "token missing or invalid" });
+  const user = request.user;
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id)
+    return response.status(401).json({ error: "token missing or invalid" });
+  const userId = request.user.id;
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else response.status(403).end();
 });
 
 blogsRouter.get("/:id", async (request, response) => {
